workers/ephemeral/python-sandbox/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "python-sandbox"
version = "0.1.0"
description = "Python sandbox for ephemeral worker"
requires-python = ">=3.10"
dependencies = [
    "backoff>=2.2.1",
    "grpcio>=1.59.0",
    "plotly>=5.19.0",
    "protobuf>=5.26.0",
    "psutil>=5.9.8",
    "pydash>=6.0.2",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "simplejson>=3.17.6",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.82.0",
]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
